gcc-config: make wrapping of programs dynamic
Rather than hand maintain a list of which programs to wrap, automatically
wrap everything found in the gcc path.  Similarly, use the old profile to
automatically delete all the old progs that we wrapped.

There is the possibility that if the old gcc profile is removed before we
switch to the new one, the exact list of programs that we wrapped will be
lost, but that shouldn't generally happen.  If it does, we'll try and fix
the case which caused it to happen instead.

URL: http://bugs.gentoo.org/238984
Signed-off-by: Mike Frysinger <vapier@gentoo.org>