gcc-config: for Prefix split-/usr is never an issue
Don't consider linking, even if /usr would be split, on Prefix.  Do
cleanup from older gcc-config though.  This leaves a gap where we don't
cleanup if the Prefix install uses a different filesystem for its /usr,
but that should be pretty rare, non-standard and utterly weird, so we
don't care.

Signed-off-by: Fabian Groffen <grobian@gentoo.org>
Signed-off-by: Mike Frysinger <vapier@gentoo.org>