gcc-config: handle offload/accel compilers properly
accel/offload compilers are named like 'x86_64-pc-linux-gnu-accel-nvptx-none-gcc'.

When building the offload compiler, we'll have ${CHOST}-${CTARGET} where
CHOST is the host and CTARGET is whatever the offload target is. gcc-config
doesn't expect this 'double tuple'.

e.g. when building nvptx-none, we'll have something like 'nvptx-none-x86_64-pc-linux-gnu-accel-nvptx-none-gcc'
and we want to strip that to be just x86_64-pc-linux-gnu-accel-nvptx-none-gcc, as that's
what we want symlinks to be made for in /usr/bin.

Bug: https://bugs.gentoo.org/945296
Signed-off-by: Sam James <sam@gentoo.org>